Moved code related to admin views into views.py
