+++ admin/__init__.py
@@ -1,46 +1,8 @@
-from flask_admin import Admin, AdminIndexView
+from flask_admin import Admin
 from auth.models import db, User
 from books.models import Book, Subject, Faculty
-from flask_admin.contrib.sqla import ModelView
 from werkzeug.security import generate_password_hash
-from flask_login import current_user
-
-
-
-class AdminIndex(AdminIndexView):
-    def is_accessible(self):
-        return current_user.is_authenticated and current_user.is_admin
-    
-    def can_edit(self):
-        return True
-    
-    def can_delete(self):
-        return True
-
-
-class AdminModelView(ModelView):
-    def is_accessible(self):
-        return current_user.is_authenticated and current_user.is_admin
-    
-    def can_edit(self):
-        return True 
-    
-    def can_delete(self):
-        return True
-    
-
-class AdminBookView(AdminModelView):
-    column_list = ['name','price','stock','desc','pub_date','faculty','subject']
-    form_columns = ['name','price','stock','desc','pub_date','faculty','subject','image']
-    column_filters = ['price','faculty','subject']
-    column_searchable_list = ['name','desc']
-
-
-class AdminUserView(AdminModelView):
-    column_list = ['id','email','username','student_id']
-    form_columns = ['email','username','student_id','password']
-    column_filters = ['email','username','student_id']
-    column_searchable_list = ['email','username','student_id']
+from .views import AdminIndex, AdminBookView, AdminUserView, AdminModelView
 
 
 #  initialize Admin instance
